% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_p_value.r
\name{apa_p_value}
\alias{apa_p_value}
\title{apa_p_value}
\usage{
apa_p_value(p)
}
\arguments{
\item{p}{p value to be formatted}
}
\description{
This function uses rounds p values using APA rules: No leading zero, rounded to three decimal places, below this shows as "< .001".
}
\examples{
apa_p_value(0.00000004)
apa_p_value(0.487173)
}
