% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFinterpretation.r
\name{BFinterpretation}
\alias{BFinterpretation}
\title{BFinterpretation}
\usage{
BFinterpretation(BF)
}
\arguments{
\item{BF}{Bayes Factor to be interpreted.}
}
\description{
This function uses Wagenmakers' (2011) interpretation of Bayes Factors for the purpose of Bayesian hypothesis testing
}
\examples{
BFinterpretation(10.7)

}
