% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.r
\name{center_vars}
\alias{center_vars}
\title{center variables}
\usage{
center_vars(x)
}
\arguments{
\item{x}{Variable or dataframe to be centered.}
}
\description{
This function allows you to combine multiple ggplot2 plots into one.
Code from https://hlplab.wordpress.com/2009/04/27/centering-several-variables/
}
\examples{
center_vars(my_data)
}
