% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cles.r
\name{cles}
\alias{cles}
\title{cles}
\usage{
cles(variable, group, data, value1 = 1, value2 = 0, runs = 10000)
}
\arguments{
\item{variable}{continuous variable}

\item{group}{dichotomous group}

\item{data}{data}

\item{value1}{assignement of group 1}

\item{value2}{assignement of group 2}

\item{runs}{max number of permutations}
}
\description{
This function calculates a permutation test version of the common language effect size. Unlike McGraw & Wong's original method, which used an algebraic/normal approximation, this version makes no assumptions about the distribution of the data.
Each X observation is paired to each Y observation. The proportion of XY pairs where X is higher than Y is then computed and output. Ties count as 0.5*TRUE.
Code adapted from https://janhove.github.io/reporting/2016/11/16/common-language-effect-sizes
}
\examples{
cles("Score", "Condition", data = data)
}
