% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_2x2_ezANOVA_output.r
\name{tidy_2x2_ezANOVA_output}
\alias{tidy_2x2_ezANOVA_output}
\title{tidy_2x2_ezANOVA_output}
\usage{
tidy_2x2_ezANOVA_output(model, X1, X2)
}
\arguments{
\item{model}{output of ezANOVA using a 2X2 design.}

\item{X1}{string name of IV1, as entered in ezANOVA}

\item{X2}{string name of IV2, as entered in ezANOVA}
}
\description{
This function takes the output of a 2x2 ezANOVA model, calculates partial eta squared effect sizes and their 95% CIs, and returns a text summary of the ANOVA results for the intercept, both main effects, and the interaction. Type III sum of squares are employed. Tested on entirely between groups ANOVAs, within and mixed within-between designs not tested.
}
\examples{
results <- tidy_2x2_ezANOVA_output(model, X1 = "IV_1", X2 = "IV_2")
}
