% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesavers.r
\docType{package}
\name{timesavers-package}
\alias{timesavers}
\alias{timesavers-package}
\title{Timesavers}
\description{
Misc functions for data processing, analysis and reporting.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@ugent.be}

}
